+++ modules/outpatient/outpatient-api/src/main/java/hk/org/hkbh/cms/outpatient/service/CodeDetailLocalServiceUtil.java
@@ -60,7 +60,7 @@ public class CodeDetailLocalServiceUtil {
 	* @return the new code detail
 	*/
 	public static hk.org.hkbh.cms.outpatient.model.CodeDetail createCodeDetail(
-		long id) {
+		int id) {
 		return getService().createCodeDetail(id);
 	}
 
@@ -83,7 +83,7 @@ public class CodeDetailLocalServiceUtil {
 	* @throws PortalException if a code detail with the primary key could not be found
 	*/
 	public static hk.org.hkbh.cms.outpatient.model.CodeDetail deleteCodeDetail(
-		long id) throws com.liferay.portal.kernel.exception.PortalException {
+		int id) throws com.liferay.portal.kernel.exception.PortalException {
 		return getService().deleteCodeDetail(id);
 	}
 
@@ -175,7 +175,7 @@ public class CodeDetailLocalServiceUtil {
 	}
 
 	public static hk.org.hkbh.cms.outpatient.model.CodeDetail fetchCodeDetail(
-		long id) {
+		int id) {
 		return getService().fetchCodeDetail(id);
 	}
 
@@ -191,7 +191,7 @@ public class CodeDetailLocalServiceUtil {
 	* @throws PortalException if a code detail with the primary key could not be found
 	*/
 	public static hk.org.hkbh.cms.outpatient.model.CodeDetail getCodeDetail(
-		long id) throws com.liferay.portal.kernel.exception.PortalException {
+		int id) throws com.liferay.portal.kernel.exception.PortalException {
 		return getService().getCodeDetail(id);
 	}
